# swagger.yaml
openapi: 3.0.3
swagger: 2.0
info:
  title: vieritage
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost:5000
components:
 securitySchemes:
  bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/v1/shorts/disupvote/{id}:
    post:
      security:
         - bearerAuth: []
      tags:
        - shorts
      summary: Disupvote
      description: Disupvote
      operationId: disupvote
      responses:
        '200':
          description: ''
  /api/v1/shorts/downvote/{id}:
    post:
      security:
         - bearerAuth: []
      tags:
        - shorts
      summary: Downvote
      description: Downvote
      operationId: downvote
      responses:
        '200':
          description: ''
  /api/v1/shorts/disdownvote/{id}:
    post:
      security:
         - bearerAuth: []
      tags:
        - shorts
      summary: Disdownvote
      description: Disdownvote
      operationId: disdownvote
      responses:
        '200':
          description: ''
...
  /api/v1/shorts/scrutinize/{id}:
    post:
      tags:
        - shorts
      summary: Scrutinize
      description: Scrutinize
      operationId: scrutinize
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  example: refuse
            example:
              action: refuse
      responses:
        '200':
          description: ''
...
  /api/v1/posts:
    get:
      security:
         - bearerAuth: []
      tags:
        - post
      summary: Get all posts of user
      description: Get all posts of user
      operationId: getAllPostsOfUser
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            example: 646ef3637251a0220e25132a
      responses:
        '200':
          description: ''
  /api/v1/posts/{id}:
    get:
      security:
         - bearerAuth: []
      tags:
        - post
      summary: Get post
      description: Get post
      operationId: getPost
      parameters:
        - name: a
          in: query
          schema:
            type: string
            example: ''
      responses:
        '200':
          description: ''
  /api/v1/posts/upload:
    post:
      security:
         - bearerAuth: []
      tags:
        - post
      summary: Upload post
      description: Upload post
      operationId: uploadPost
      parameters:
        - name: medias
          in: header
          schema:
            type: string
            example: ''
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: abc
                description:
                  type: string
                  example: def
                genres[]:
                  type: string
                  example: '["6470048f669d752ca5b40e94", "647004bd669d752ca5b40e96"]'
                medias:
                  type: string
                  format: binary
      responses:
        '200':
          description: ''
...